fix: types

+++ packages/contracts-bedrock/interfaces/L1/IFeesDepositor.sol
@@ -9,7 +9,7 @@ import { IOptimismPortal2 as IOptimismPortal } from "interfaces/L1/IOptimismPort
 interface IFeesDepositor is ISemver, IProxyAdminOwnedBase, IReinitializableBase {
     event FundsReceived(address indexed sender, uint256 amount, uint256 newBalance);
     event FeesDeposited(address indexed l2Recipient, uint256 amount);
-    event MinDepositAmountUpdated(uint256 oldminDepositAmount, uint256 newminDepositAmount);
+    event MinDepositAmountUpdated(uint96 oldminDepositAmount, uint96 newminDepositAmount);
     event L2RecipientUpdated(address oldL2Recipient, address newL2Recipient);
     event GasLimitUpdated(uint64 oldGasLimit, uint64 newGasLimit);
     event DepositDataUpdated(bytes oldDepositData, bytes newDepositData);
@@ -20,7 +20,7 @@ interface IFeesDepositor is ISemver, IProxyAdminOwnedBase, IReinitializableBase 
     function gasLimit() external view returns (uint64);
     function depositData() external view returns (bytes memory);
     function initialize(
-        uint256 _minDepositAmount,
+        uint96 _minDepositAmount,
         address _l2Recipient,
         IOptimismPortal _portal,
         uint64 _gasLimit,
@@ -28,7 +28,7 @@ interface IFeesDepositor is ISemver, IProxyAdminOwnedBase, IReinitializableBase 
     )
         external;
 
-    function setMinDepositAmount(uint256 _minDepositAmount) external;
+    function setMinDepositAmount(uint96 _minDepositAmount) external;
     function setL2Recipient(address _l2Recipient) external;
     function setGasLimit(uint64 _gasLimit) external;
     function setDepositData(bytes memory _depositData) external;
